remove notification options
Their introduction was a mistake in the first place.
It makes no sense to silently drop fails/errors.
It *may* make sense to silently drop OKs but in the end
users won't know whether everything was OK or an error along the way
caused radio silence.
Removal will make this extension harder to misuse.